handle writeItems request with >50 items
transparently break array into chunks of 50 and update statuses
writeItems calls should look the same to consumers regardless of how many
items they pass in.
It also no longer returns False on a request failure. Instead of updates the
writeFailure for each item with the code and message of the response.