python items.deleteItem and items.deleteItems
deleteItems currently will not accept more than 50 items
To do that we'd need to deal with returning the responses from multiple requests
which may have different results. For the API, a single batch delete is atomic,
so we could return bool for these instead of the http response, but for deletes
requiring multiple requests we'd need to return multiple results.